driver: include the original exception if fails to load
When a driver tries to load one extension and fails to do so, the
current message is coming from _init_plugins() that states that no
extension were found. Otherwise, that might be wrong. The problem could
be that invoke_on_load=True was set and that the driver was unable to be
instanciated. In that case, it's much better to include this exception
string into the original exception, so it eases the debugging.

This fixes #32